Change name of main index.js
